chore: move input-password in separate example
