+++ projects/demo/src/modules/components/textfield/examples/4/index.html
@@ -1,17 +1,4 @@
 <tui-textfield>
-    <label tuiLabel>Click icon to copy</label>
-    <input
-        placeholder="I am placeholder"
-        tuiTextfield
-        [(ngModel)]="value"
-    />
-    <tui-icon
-        icon="@tui.copy"
-        tuiCopy
-    />
-</tui-textfield>
-
-<tui-textfield>
     <label tuiLabel>Click icon to toggle</label>
     <input
         placeholder="I am placeholder"
