--- projects/demo/src/modules/components/textfield/examples/4/index.less
@@ -1,5 +1,0 @@
-:host {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
