+++ projects/demo/src/modules/components/textfield/index.html
@@ -40,8 +40,12 @@
                         ComboBox
                     </a>
                 </span>
-                <span *ngSwitchCase="3">Example of input-password and input-copy with interactive icons</span>
-                <div *ngSwitchCase="4">
+                <div *ngSwitchCase="3">
+                    An example of a password textfield with an interactive icon to toggle the input type.
+                </div>
+                <div *ngSwitchCase="4">An example of a textfield with an interactive icon to copy the input value.</div>
+
+                <div *ngSwitchCase="5">
                     We recommends using
                     <a
                         href="https://maskito.dev"
